server/requirements.txt
pandas
numpy
flask
openai
python-dotenv
requests
pytesseract
opencv-python
icalendar
datetime
pathlib